Cache clients created from a call to useClients
When a programmer passes CreateClientArgs to useClients, graz will create a
client for the given chain, but will not cache it in the graz store, leading
to strange issues where a client has been connected but cannot be retrieved
via useClients calls with no arguments. This change stores the clients in the
zustand store to avoid such issues.